chore: move admin generation to `user.seed.ts`
