+++ src/server/db/seed.ts
@@ -8,24 +8,11 @@ import {
   cities,
   countries,
   languages,
-  profiles,
-  profilesAddresses,
-  profileSkills,
-  profilesLanguages,
-  profilesLicences,
-  Sex,
   users,
 } from "../db/schema";
-import { type PgTransaction } from "drizzle-orm/pg-core";
 import { populateEducations } from "./seed/education.seed";
 import { populateLicenses } from "./seed/license.seed";
-import {
-  generateProfileSkills,
-  generateUsers,
-  insertWorkStatus,
-  SALT_OR_ROUNDS,
-  getRandomLanguages,
-} from "./seed/user.seed";
+import { generateUsers, SALT_OR_ROUNDS, generateAdmin } from "./seed/user.seed";
 
 const generateCountriesWithCities = async (): Promise<string[]> => {
   const _cities = [
@@ -153,60 +140,13 @@ const main = async () => {
     where: eq(users.email, email),
   });
   if (!adminExists) {
-    await db.transaction(async (tx) => {
-      const [admin] = await tx
-        .insert(users)
-        .values({
-          email,
-          password: adminPassword,
-          active: true,
-        })
-        .returning();
-
-      if (!admin) throw new Error("Admin user could not be created");
-
-      const [adminProfile] = await tx
-        .insert(profiles)
-        .values({
-          userId: admin?.id,
-          firstName: "Admin",
-          lastName: "Admin",
-          oib: Math.floor(10000000000 + Math.random() * 90000000000).toString(),
-          sex: Sex.MALE,
-          nationality: "Foo",
-          parentName: "Test",
-          birthDate: new Date(1990, 1, 1).toISOString(),
-        })
-        .returning();
-
-      if (adminProfile && _licences.length) {
-        await tx.insert(profilesLicences).values({
-          profileId: adminProfile.id,
-          licenceId:
-            _licences[Math.floor(Math.random() * _licences.length)]?.id ??
-            _licences[0]!.id,
-        });
-
-        const _profileSkills = generateProfileSkills(adminProfile.id);
-        await tx.insert(profileSkills).values(_profileSkills);
-
-        await tx
-          .insert(profilesLanguages)
-          .values(getRandomLanguages(adminProfile.id, languageIds));
-
-        await tx.insert(profilesAddresses).values({
-          profileId: adminProfile.id,
-          addressId: addressIds[Math.floor(Math.random() * addressIds.length)]!,
-          isPrimary: true,
-        });
-
-        await insertWorkStatus(
-          // eslint-disable-next-line @typescript-eslint/no-explicit-any
-          tx as unknown as PgTransaction<any>,
-          adminProfile.id,
-        );
-      }
-    });
+    await generateAdmin(
+      email,
+      adminPassword,
+      _licences,
+      languageIds,
+      addressIds,
+    );
   }
 
   await generateUsers(addressIds, languageIds, _licences);
